actualizado_el_2022_03_13

+++ examen2Parc/src/ejerc3/MiTriangulo.java
@@ -4,11 +4,11 @@ import base.LadosIlegalesException;
 import base.Triangulo;
 
 public class MiTriangulo extends Triangulo {
-	private MiTriangulo(int a, int b, int c) {
+	private MiTriangulo(double a, double b, double c) {
 		super(a, b, c);
 	}
 
-	public static MiTriangulo equilatero(int a) {
+	public static MiTriangulo equilatero(double a) {
 		return escaleno(a, a, a);
 //		try {
 //			return new MiTriangulo(a, a, a);
@@ -16,7 +16,7 @@ public class MiTriangulo extends Triangulo {
 //			return null;
 //		}
 	}
-	public static MiTriangulo isosceles(int a, int b) {
+	public static MiTriangulo isosceles(double a, double b) {
 		return escaleno(a, a, b);
 //		try {
 //			return new MiTriangulo(a, a, b);
@@ -24,7 +24,7 @@ public class MiTriangulo extends Triangulo {
 //			return null;
 //		}
 	}
-	public static MiTriangulo escaleno(int a, int b, int c) {
+	public static MiTriangulo escaleno(double a, double b, double c) {
 		try {
 			return new MiTriangulo(a, b, c);
 		} catch (LadosIlegalesException e) {
